Migrate to SPDX license
This is a mess. There is no sign of BSD - as far as source git allowed me to go. The frontpage at Savanah claims that the license is X11. But the license file and file headers refer to plain MIT.
This is part of https://fedoraproject.org/wiki/Changes/SPDX_Licenses_Phase_2